Move fileserver tests into handlers
Follow the same test layout pattern of reflecting the layout of the
source code under test and move all the fileserver tests into the
tests/handlers directory. Put all the tests in the same file, remove
a few unnecessary bits, and tighten up the code a bit.
